customizable SuggestionsList (#439)
* customizable SuggestionsList

* fix examples

* fix bug

* configurable SuggestionsListItem
