+++ src/components/SuggestionsList/SuggestionsListItem.tsx
@@ -1,27 +1,20 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 import '../../yamui';
 import * as React from 'react';
+import { CustomizableComponentProps, defaultTheme, customizable } from '../Customizer';
 import { join } from '../../util/classNames';
-import { BaseComponentProps } from '../../util/BaseComponent/props';
 import { getClassNames } from './SuggestionsListItem.styles';
 import SuggestionsListItemContent from './SuggestionsListItemContent';
-import { SuggestionItem } from './types';
-
-export interface SuggestionsListItemProps extends BaseComponentProps {
-  item: SuggestionItem;
-  searchText: string;
-  isSelected: boolean;
-  onSelect(id: SuggestionItem['id']): void;
-}
+import { SuggestionsListItemProps } from './SuggestionsList.types';
 
 const baseClass = 'y-suggestionsListItem y-hc-select-on-hover y-hc-suppress-text-background';
 const selectedClass = `${baseClass} y-hc-selected`;
 
-export default class SuggestionsListItem extends React.PureComponent<SuggestionsListItemProps> {
+export class SuggestionsListItem extends React.PureComponent<SuggestionsListItemProps & CustomizableComponentProps> {
   public render() {
-    const { className, isSelected, item, searchText } = this.props;
+    const { className, isSelected, item, searchText, theme = defaultTheme } = this.props;
     const baseClassNames = isSelected ? selectedClass : baseClass;
-    const classNames = getClassNames({ isSelected });
+    const classNames = getClassNames({ isSelected, theme });
 
     // role=button added so that speech software knows that these are clickable targets.
     return (
@@ -37,3 +30,6 @@ export default class SuggestionsListItem extends React.PureComponent<Suggestions
     onSelect(item.id);
   };
 }
+
+@customizable('SuggestionsListItem', ['theme'])
+export default class CustomizableSuggestionsListItem extends SuggestionsListItem {}
